Refactor : 무한스크롤 게시글 분리
1. 메인페이지 외 다른 페이지에서도 무한스크롤 게시글을 사용해야하는 경우가 있어서 분리했습니다.

InfinityPost.tsx props로 API_URL만 내려주시면 되고 엔드포인트 이후 값 넣어주시면 됩니다.

2. post.type.ts 변수 명 Post[] -> IPost[]로 변경했습니다. 인터페이스 참조하시는 분은 변경 부탁드립니다.
